Updates Sentinel buildHydratedDoc to treat the case of empty lineage
When processing changes, fetchHydratedDoc is called before any
transition. In the case of SMS forms, for example, buildHydratedDoc will
receive an empty array as lineage parameter, which will cause it to
throw a TypeError since doc.contact will be undefined.

The fix is to return original document if lineage array is empty.