Refactoring fixes
Reverted the exportDataV1 call to use callbacks as the code now does.

Moved the flushHeaders into V2 only -- if you flush early you need to
be careful to not set headers again.

Changed V2 code to support only flushing headers once. Now we abort the
response with an error code in the output.

medic/medic-webapp#3826