Simplify travis build output
Make the travis build log quicker to load and easier to understand at a glance.

This also changes previous behaviour so that uploads to s3 of test reports only happen for failed build jobs, rather than all.

An example of a successful and failed build can be seen at https://travis-ci.org/medic/medic-webapp/builds/356822011.

Log files are suffixed with .txt so that they are served with a suitable content-type header by s3 for direct reading in the browser.

Closes #4316
Closes #4317