Reports verification update
Updates the UI to show an additional actionbar when choosing to
verify a report, which contains 2 options to convey whether the
report is valid or not.
Verification validity is saved as `verified_valid` (bool) property in
the doc.
Updates `doc_summaries_by_id` to include this property in the summary.
Adds new permission `can_verify_reports` for displaying reports
verification button.