Implements configurable roles
Makes roles configurable so instances can define different groups of users to assign
permissions to, for example: "supervisor", "chp", "innovation-network", etc.

Each role can be set as online or offline to indicate whether the db should be replicated
locally or not.

The user doc in the _users db and the user-settings doc in the medic database both
have an array of roles as this is consistent with CouchDB, but the configuration UI has a
dropdown so in practice users will have an array of length 1.

The default configuration is backwards compatible with the old hardcoded roles so no
migration is necessary.

medic/medic-webapp#4525