Date last visited support
Add a new permission that allows for date last visited to be displayed on the
person list, and for sorting by date last visited here as well.

This is currently experimental and will probably not scale to larger lists
of people. It is also hard coded to only check for health centres, which are
usually considered "families" by many deploys, which is enough to A/B test
this feature for now.

medic/medic-webapp#4524
medic/medic-webapp#4526