Adds the MRDT enketo widget and android integration
This adds a new widget for MRDT verification which launches a third party app which
returns a photo of the MRDT result which gets attached to the report and displayed in
the report view.

Attaching binary content to the report is done generically so other widgets can use the
same pattern.

medic/medic-webapp#4743
medic/medic-webapp#4745
medic/medic-webapp#4742