Process CouchDB delete-stubs correctly in LiveList and TombstoneUtils
When a doc is deleted via a DELETE call (or from Fauxton UI), CouchDB
only saves a stubbed version of the doc (containing just _id, _rev and
_deleted: true flag).

When such a change was received by webapp's listeners, it was not
filtered as relevant, since the attached doc didn't have the required
properties, so the LiveList was not updated. This modifies the webapp
changes listeners to filter delete stubs of docs that exist in the
LiveList.

Updates tombstone-utils shared library so it will try to save the
previous revision of delete-stubs as the tombstone, so offline users
will replicate the DELETEs.

Additionally, support restricted offline users, namely:
Allow db-doc GET and _bulk_get to read delete stubs.

medic/medic-webapp#4754