Update README.md (#4781)
Adding some sections about browser support and a very quick summary of
medic-conf. I would expect that any real tutorial for that would be in
its own documentation.

I also dropped the reference to Medic OS. We can add that back once we
have correct instructions for Docker to be run on desktops.