LiveList full refresh (#4785)
Adds support for truly refreshing the LiveList.
This is achieved by providing a list of currently loaded doc ids to
webapp's Search service.
After querying as usual, the results are merged with the provided
doc ids prior to calling GetDataRecords and other operations,
ensuring that all existent items from the LiveList are updated.
This mechanism is only triggered for Contacts when sorting is by
last_visited_date and change with a visit report about an loaded
contact is received.

medic/medic-webapp#4782