Fixes search debouncing by cloning filter objects
Any change to report filters are ignored while reports are being fetched. The function `debounce()` in `services/search.js` is effectively bouncing everything because `_currentQuery.filters` is a reference to `$scope.filters` from `controllers/reports.js`. Fix is to store a copy of the objects instead of a reference.

medic/medic-webapp#4331