Adds muting history
A contact's muting_history log is saved in the sentinel info doc.
Each time the muted status is changed, an entry is added in the
muting_history log.
A contact's muted property now contains a human readable date
(in ISO8601 format). When unmuting, the property is removed altogether,
avoiding having mixed data types for the same property.
The muted field value (when existent) is exposed to ContactViewModel,
instead of a Boolean true.

#4767