Support binding to fields.patient_uuid
In some configurations the patient_id field holds the patient shortcode,
while the patient_uuid field holds the internal id.

For tasks and targets it's important that we know how to bind to the
UUID as well, because otherwise users may see bugs while offline. For
example, a task wouldn't close until the report that closes the task is
synced upwards, run through by Sentinel, and synced back down.

Less important for purging, but it's better that we keep the function
logic the same.