Serves new appcache from old location
When users upgrade from 2.x their old appcache needs to be
updated with new files. This makes sure the appcache is
available at the old location as well as the new location.

This is not a long term solution but will do until we
migrate to using service worker caching instead.

medic/medic-webapp#5247