Removing use of data_records_by_type query
NB: in production we have to have two strategies, one for offline (as
shown here) and one for online users: probably keeping the mapreduce I
just used (or using a more generic view like docs_by_type and adding a
reduce to it, which would remove the need for a completely separate
view).

This trick is fine for offline users because the records have to be in
memory anyway.