Included Docker instructions (#5333)
Included Docker instructions, and removed horti instructions since it wasn't clear why someone would use it. Moved the horti instructions to https://github.com/medic/medic-docs/blob/master/installation/deploy-locally-using-horticulturalist.md for posterity. If there is a good reason to include it in the readme we can link to it.

Also converted mentions of medic-os to Docker