Drop optimise event, move start event
Drop "optimise", the manual compaction step. We don't need to do this
as we are compacting on write as we go.

Move start. This now happens once we decide we are going to purge, as
it's more useful to show this early and have something to show while
we pull in and organise documents that need purging.

medic/medic#5348