Fixes navigation warning popup
Fixes transition bug when when editing forms and navigating to other
tabs.
Changes Popup behavior when pressing "YES" so it directs you to the
desired URL instead of always treating it as the "cancelCallback".
Does not display popup when editing a form and executing a search
(contacts & reports).
Fixes popup displayed after task submission.

medic/medic#5409