Slightly optimise how we store our log history
Instead of constantly onto the front of the array and pruning by popping
off the end, instead keep a rolling index into the space and leave the
ordering problem to when someone requests the log.

This gives a constant 10% speed boost in writing logs (for me in V8),
though this is probably pretty JS engine specific.