Moves the publish step into CI script (#5586)
Apparently errors in the `after_success` section don't fail the build. This causes the subsequent jobs to fail with a misleading error ("failed to start API"). Moving it into the `script` section causes the `compile` job to fail so it fails earlier and in a way that's easier to debug.

medic/medic#5196