Alleviate performance issue related to report tab
With this commit, the report tab will load without the facilities tree filter being built. When the user clicks on the tab, an initial load of the facilities tree takes place. Subsequent load happen as the user is scrolling through the list of facilities hierarchy and reaching the bottom of the list.