Include language for CHT (#5603)
Updated intro to include CH App Framework

There are many changes needed throughout the repo if we were to switch everything to CHT language. The purpose of this PR is to bring in language for the transitional state so that it makes sense with were we are at now -- having a medic repo as the Community Health Application Framework.