New permission 'can_edit_verification' blocks changing of report verification state once set (#5575)
Partners (like Muso) who are looking to capture analytics data on time intervals involving report verification have requested for users to be unable to change the verification state of a report once set. This change adds a new user-level permission "can_edit_verification" to achieve this. When false, users are prompted to confirm the state before it is set.