Set initial_creation_date of infodocs to when they are first created
This is not totally correct but is a less destabilising solution than a
more correct one.

For new documents the date will be off by however long it takes for
sentinel to catch up (which could be <1s).

For old documents that don't already have info docs the date will be
hilariously off.

A more correct solution would be to keep this code as is, and to create
infodocs in API when documents get PUT/POST/bulk_docs'ed

medic/medic#5760