Ensure RulesEngine.init warms the contact_by_type view (#5786)
The load time of the contacts tab increased from 3.4 to 3.5. This regression is actually only in the "first-load time", and does not impact subsequent loads. It is caused because the contact_by_type index is no-longer warmed during bootstrap.

By setting the default query of Search to be contact_by_type, this change ensures that the contact_by_type view is once-again warmed during bootstrap. It seems that the contact_by_type_index_name is then no longer used.

Did the best I can to test for regressions, but there are a lot of scenarios for the Search service. Would welcome your keen insights, @dianabarsan. Did not measure impact on bootstrap times.

#5755