Contact controller should load children/reports in parallel
This is a minor perf improvement for how we load contacts. Prior to the redux refactor, the ContactViewModelGenerator handled this in parallel with the logic in the ContactsController. Now the ContractsController handles loading this in parallel.

Before	After
1286	1139
1196	1150
1287	1143
1150	1078
1143	1113
 	 
1212.4	1124.6
#5755

I've also never worked with redux + angular together before. I'm attempting to clean up a few things, but if I'm breaking any patterns here, please let me know.

I personally find performance investigations quite difficult in our product. Part of this, is because profiles are hard to read on AngularJS. Another part is that many layers of the code (services, controllers, actions, shared-libs) can all query the database. I have worked on products previously that had the concept of a "data access layer" (or Repository Pattern), where only one specific layer is allowed to fetch data from the database. This refactoring makes it so at least actions are not querying the database. @rmhowe I'm wondering what you think of this suggestion?

Also, we previously had two actions which we would almost never call separately. One to say we are loading a contact's children, and one to say we are loading a contact's reports. My understanding of actions is that they should be mapped from interactions with the UI - and so it should be safe to combine these.