Allow mRDT Supervisor Tasks to resolve based on multiple needs_signoff reports (#5790)
Because of replication depth, the contact_id for a task's action may not be present on the device. This change makes it so enketo does a "best effort" to inject contact data when none is present, instead of throwing when building a ContactSummary.

Reports associated with contacts which are not on the device are not grouped under the same contact even when the associated contact id is the same. This updates RulesEngine to group reports by contact id in such cases.

#5789