Rework initial security setup for CouchDB 2.3.x
Reworks the ordering to be a bit clearer.

Removes the creation of the _users DB since you should be doing that as
part of CouchDB setup (along side _replicator etc) that you either do
manually or allow their wizard to do for you.

Fixes the use of credentials for CouchDB 2.3.0: CouchDB is no longer in
admin party, and so you create DBs or alter config with some kind of
credential.