Warn offline users before replicating too many docs (#5793)
Adds /api/v1/users-info endpoint which

returns the number of docs a user will replicate and a warn flag if this number exceeds the recommended limit.
when queried as an offline user, returns the requester doc count.
when queried as an online user, requires facility_id and role params, plus an optional contact_id param (either GET or POST).
is queried as a new step in webapp bootstrapper. When resulting warn flag is truthy, the user has to confirm to actually continue bootstrapping and start replication. Cancelling redirects to login.
is queried as a new step in admin user creation/updating. When resulting warn flag is truthy, a warning message is displayed. The admin would need to click on "Submit" the 2nd time to actually create or update the user document.
Once an hour, API will log warnings of users that have replicated more than 10 000 docs.
Updates webapp bootstrapper to display correct doc count progression, even when interrupted.

#5362