Add first release test - Please DO NOT MERGE (#5908)
Refactored to no longer use *.conf.js and control through the grunt execution. Which is replaced by a conf.js similar to the protractor default.
Added suites to differentiate which tests are being ran. 
Added applyconfig to apply app configs for scenarios based on a configuration of the app
Added first release test.