Server side form generation
Generate the "model.xml" and "form.html" files on the server.

This simplifies and improves performance on the client. The first load will be quicker as
no xslt stage is required. In addition we no longer need to cache the result so memory
usage is reduced.

This also means we can use an (almost) stock enketo-xslt which makes it much easier to
keep on the latest version of enketo.

It also introduces a new runtime dependency on the xsltproc command.

#4450