Server side purge (#5878)
Adds serverside pseudo purge which runs on a schedule and executes a configurable "purge" function over all docs that would be replicated. "Purged" docs are added in auxiliary databases - one for every unique user roles group.
During offline user replication (both initial and subsequent), only docs that are not tagged as "purged" are downloaded. On a schedule, on application bootstrap, the webapp will check if new documents were "purged" and delete them from the device.

#5443