CHT Reference App (#5927)
The idea of a reference app was born of two truths: 1) designing and building an app from scratch is a significant undertaking, and 2) many unique program models exist and there is no such thing as a “one size fits all” app.

A reference app provides a foundation for forms, data fields, and analytics so that partners can easily customize workflows to their context. It has been designed to model an ideal way of structuring and organizing a workflow, writing configuration code, and authoring tests.