5914 - Use CHT Reference App as default configuration (#5945)
If the CouchDB database doesn't contain any configuration, deploys a default configuration. This change updates the grunt build process to use medic-conf@3.0.5 to package the cht-reference-app configuration (config/default folder) into files which are ready to be consumed by couchdb. When API starts, if no settings document exists when the CouchDB database starts, the cht-reference-app documents are pushed.