Updating instructions
Reworking our easy deployment section to make it even easier
(hopefully), by adding GUI instructions (kitematic) and a
docker-compose file for CLI, for folks who don't want to do core
development and instead are just checking out the reference app or are
developing configurations of their own.

Moved the Supported Browsers section up as it's not realy anything to do
with development setup, and should really be known by everyone.