Add a timeout to outbound requests
There have been instances where the outbound request handler has hung
completely in production. It's not clear from logging why this occured
and it's not consistent or repeatable, but a good potential option is
that requests to external servers are hanging.

This change adds an explicit timeout to the two outbound IO requests.

Depending on various factors there may already be an implicit timeout,
but this makes it explicit and much shorter (10 seconds).

medic/cht-core#6025