Make application versions consistent
This ensures the two version numbers (the git tag, and the npm
package.json) are equivalent. This is important to avoid bugs
where the version is incorrectly reported.

The two changes are:
1. add a travis build check that if building a tag, the tag name
starts with the package.json name, and
2. remove the version from the logo tooltip as it was seldom seen
and not as accurate as the version on the About page

This also triggered a range of refactoring to clean up the version
code.

medic/cht-core#6054