Update config/standard and config/default to use medic-conf 3.1
* Update config/standard to use medic-conf 3.1 (#6156)

This updates the standard config to be ready for cht-core v3.8 release and upgrade to medic-conf 3.1. medic/medic-conf#275. This is required to use the standard configuration with projects running webapp 3.8's rules-engine v2.

* Update config/default to use medic-conf 3.1 (#6155)

This updates the default config to be ready for cht-core v3.8 release and upgrade to medic-conf 3.1. medic/medic-conf#275. This is required to use the default configuration with projects running webapp 3.8's rules-engine v2.