Creating a npm-level everything action
doing this inside grunt doesn't work because of how our application is
architected. Specifically, places that look for UNIT_TEST_ENV only do so
once on initialisation, and grunt doesn't restart the VM between
actions.

Instead, if we run this from npm it just executes this like bash would,
with fresh vms between each run (which is way better regardless of this
issue)