Adds self_report transition. (#6377)
Adds a new transition called self_report. Configuration accepts a list of forms that it runs on and a list of messages to send, based on events. 
Valid documents are data_records that have a "sender" (from field), are configured to use the transition and don't have a patient (even if the patient is not found).
Populates fields.patient_id and fields.patient_uuid with data from the sender.
The transition is second to run over the doc (after update_clinics) so it provides the hydrated patient to all subsequent transitions.

#6286