Outbound will send multiple requests for the same record
Adds a hash into completed_tasks so we can correctly detect when to
re-send a record.

This changes the sending semantics for outbound to be that we send an
outbound request whenever we have a configuration with a truthy
relevant_to, and where the prior send (if any) for that configuration
sent a different payload.

This is not a perfect system. If your mapping function is not idempotent
outbound may fire requests when you do not intend it to, e.g., if a user
doesn't actually edit the document meaningfully, or if sentinel or
another process writes to the document.

medic/cht-core#6419