Adds configurable header tabs icons (#6499)
Adds the header_tabs section in app_settings that can be used to configure which icons can be used by the app header tabs. Icons can be resource icons or FontAwesome icons, with resources taking precedence.
Changes resourceIcon filter to accept a FontAwesome placeholder that is displayed if the resource icon is not found.
Adds a new tab under the Images page in the Admin Console where header tabs icons can be configured. Resource icons are filtered to only display svgs.

#6410