Single emissions recalculation thread (#6566)
Updates rules-engine shared lib to only allow a single refreshRulesEmissions call to be executed in parallel. All calls to fetchTasksFor and fetchTargets are queued and are execute sequentially.
To support accurate Telemetry readings, calls to fetchTargets and fetchTasksFor return promisemitters, which resolve when the tasks are done and emit two events: queued when placed in the queue and running when execution started.
The webapp RulesEngine service listens to these events and logs a new Telemetry record, queued, that registers the time the request spent in a queue (from queued event to running event), along with the already existent "execution" record, that only registers execution time (from the running event to promise resolution).

#6562

(cherry picked from commit fb4b8cc0314570febaa4e0c9498a8ca66878e4bf)