Use environment variable to set webdriver version (#6620)
Instead of committing the version in the repo, this change uses an environment
variable, which means it's much easier to change when we need to. The value
also has a default so it's easy to run in development. Travis has been configured
to provide the environment variable so we can bump it for all builds from one place.