Adds translations compiler to support pluralization with message format (#6715)
We used to rely on messageformat to provide pluralization. The AngularJS translator allowed you to control how you could interpolate a translation, but ngx-translate does not.
Because we only want to pass "some" translated strings to message-format, I leveraging the option of adding a compiler to ngx-translate that will try to filter correct phrases to be passed to message-format.