Task completion landing page (#7234)
Adds task completion landing page, which:

- is only displayed to users that have "can_view_tasks_group" permission. If users are missing the permission, the page is not displayed. Default config is updated to grant this permission to the "chw" role.
- is only displayed after submitted a task. Routing to the URL manually or using browser history won't re-display the page.
depending on the last submitted task, displays tasks that are "Ready" that belong to the same leaf type place as the submitted task's patient. If no tasks are found or a leaf type place was not found in patient lineage, the page is not displayed.
For example (with default hierarchy), completing a task about a:
a) person that belongs to a clinic, will display all ready tasks for that clinic
b) clinic, will display all ready tasks for that clinic
c) person that belongs to a health_center, will not display the page
d) health_center, will not display the page
- if the page should be displayed (user has permission, contact belongs to or is a leaf type place), but all tasks are completed, the page is not displayed and the snackbar should display a "All household tasks completed" message.
- when the page is displayed, navigating away from the page to:
a) another random page (like contacts or reports)
b) a task that is NOT from the same leaf type place
the user should see a modal to confirm navigation, similar to the form edited popup, but with different text. 
If the user navigates to a task that is from the same leaf type place - whether the click is in the LHS or RHS list, the page should redirect to the task directly, without a modal.
- the tasks displayed on the page should be identical to the ones displayed on the contact page of the place, including the styling. 
For performance reasons, the page uses the tasks that already exists in the Store, and cherry picks which tasks to display based on the last submitted task's patient. This list can become stale between task recalculations - this behavior already exists in the tasks list. Also, the cherry picked list will exclude the last submitted task. The tasks are re-cherry picked every time the main task list is updated.
Also for performance reasons, the associated leaf type place is "preloaded" into the store by the task content component, after the form has been rendered. 

#5886