Generate service worker in API (#7322)
Removes service-worker grunt task.
Ddoc extraction no longer bumps the service-worker-meta doc.

Service worker is now generated in API:
- once on startup, after migrations
- every time the main ddoc is updated
- every time translations are changed (or added or removed)
- every time the branding doc is changed
If the call to sw-precache fails, the process crashes. Other errors are caught.
Config that we pass to sw-precache is unchanged, excepting:
- ignoreUrlParametersMatching now also includes username to fix the login page cache.
- dynamicUrlToDependencies for the login page receives the generated html of login page (that includes branding and translations), to make sure cache is busted at the correct times. As a consequence, the login controller now exports a renderLogin method that returns the generated login page.
After the service-worker file is generated (and saved in the same location as before), the service worker meta doc is updated. Updating the meta doc is skipped if no changes are made to the service-worker script (if the sha1 hash of the former file matches the sha1 hash of the new file).

Migrates service-worker e2e test to wdio.
Fixes race condition in bootstrap service-woker installation code, where registration promises only resolved when the service worker was already installed. On slow connections, where installation had not yet succeeded on the login page, the page would be left hanging.

#6338
#7307
#7242