Enables support for Bikram Sambat dates through webapp (#7342)
Adds date conversion to Bikram Sambat when selected locale is "ne". Converts:
- item list dates (like reported date in reports list, or messages) - this includes the tooltip
- last sync date
- contact summary fields that use date filters: dateOfDeath, age, dayMonth, relativeDate, relativeDay, simpleDate, simpleDateTime, fullDate

Does not convert the date filter on the reports page.
Adds a new Enketo xpath function to-bikram-sambat that converts a gregorian date to Bikram Sambat text.
Adds a new FormatNumber service that can convert digits in strings and numbers to Devanagari script. Adds a new pipe localizeNumber that calls this service, which is now used in many places where we display numbers.

#7294