Another attempt to make these tests reliable
This removes the check for the temporary message which may cause
the test to fail because it disappears too quickly. By default
the waitUntil function checks every 500ms which is likely too slow
to detect the syncing message. An alternative would be to decrease this
interval, or find a way to slow down the sync sufficiently, but if
this works it's an improvement and sufficient for now.

medic/cht-core#7580