Support replacing offline users from offline device (#7709)
Adds a new form that can be submitted from any place's primary contact page to replace that contact and its user.  
The form triggers a client-side transition that reparents all subsequent reports to the new contact.  
When the data gets sync'd, a sentinel transition runs to create that new contact's user, send a `token_login` link by SMS to the user to log into the new user account and finally log the current user out.

Co-authored-by: m5r <mokht@rmi.al>
Co-authored-by: Joshua Kuestersteffen <jkuester@kuester7.com>
Co-authored-by: Njuguna Ndung'u <njgndungu@yahoo.com>