perf(#8134): Refactor initial replication (#8141)
Adds new api endpoint that will return all necessary bits for initial replication (doc-ids-revs, warn count, whether to warn and update_seq of the medic db). 
Client iterates over doc-ids-revs in batches of 100, checking if existing locally. if not, downloads with attachments with _bulk_get
Client calls .replicate.from and .replicate.to with since parameters to generate up-to-date checkpointers to be used by subsequent syncs.

#8134