fix(#8252): Adds retry when getting userCtx when logging in (#8258)
In clustered couchdb, if a user's password is updated on one node, and immediately afterwards another node receives a login request, there's a chance it will respond with a 401 because the nodes are not in sync. This becomes very obvious when logging in with a token, where we change the user's password upon token activation, and immediately attempt a login with the new password. 
Ads a retry when getting `userCtx`, until the request either hits the same node, or enough time has passed that the nodes are already in sync. 

#8252