Adding Matomo analytics & removing others
This gets the boilerplate Matomo tag/js in. It still lacks the correct config for the MATOMO_URL and IDSITE so isn't usable until I get those values. Those of course will eventually be env vars or similar, but for now we're just hackin' this into place to see some initial records fly by.