chore(#8697): Consolidate enketo common selectors for default config (#8720)
Create a new file (common-enketo.wdio.page.js) to consolidate all the common selectors related to enketo using mainly the text label. All. the tests were updated to use that file so the other "*.wdio.page.js" files were cleaned and only contain the methods that are specific to a test.