Switch up null-checks in map function.
Trying to strike a balance here between avoiding polluting the view with bad data from bad telemetry docs vs hiding the latest telemetry data from an incomplete doc.  Without the user/date we cannot actually calculate the latest telemetry for the user. But, if we have that information, it seems reasonable to output whatever data we actually have from the doc (otherwise we could hide what we know is the latest data).