fix(#9298): no padstart in couchdb views (#9307)
Removes padStart from CouchDb views. This function was not available in the Javascript engine packaged with CouchDb v2, which ships in CHT v4.3.x and lower.
Unfortunately, when the view code crashes, it triggers a warning in CouchDb and the document is just skipped, while the view index advances. This means that that document needs to be edited in order to appear as a view result, or the whole index needs to be rebuilt, after upgrading to CouchDb 3.

#9298