chore(#9915): replace deprecated fields from `.xml` forms
This updates the `.xml` forms in `config/default`
and `config/covid-19` to replace the deprecated
`db-object`, `horizontal` and `horizontal-compact`
fields with the appropriate fields.

This commit is required because some create/edit
place `.xml` forms are not updated when we
use `cht-conf`'s `convert-contact-forms`.
Particulary the forms associated with
PLACE_TYPE-create/edit.xlsx