feat(#9301): expose user's contact summary in forms (#9824)
Adds user-contact-summary service which computes and caches the result of the contact summary of the current logged in user's contact.
This value is passed to form content, through a user-contact-summary instance, and is passed to form context, through the user.summary variable.

Fixes bug where incoming reports for relevant contacts do not show up on contact detail pages automatically.

closes #9301
closes #9825